refactor: added contract interface
