+++ app/src/main/java/com/esoxjem/movieguide/movies/details/DetailsModule.kt
@@ -1,5 +1,6 @@
 package com.esoxjem.movieguide.movies.details
 
+import com.esoxjem.movieguide.movies.favorites.FavoritesContract
 import com.esoxjem.movieguide.movies.favorites.FavoritesInteractor
 import com.esoxjem.movieguide.network.TmdbWebService
 
@@ -14,14 +15,14 @@ import dagger.Provides
 class DetailsModule {
     @Provides
     @DetailsScope
-    fun provideInteractor(tmdbWebService: TmdbWebService): MovieDetailsInteractor {
-        return MovieDetailsInteractorImpl(tmdbWebService)
+    fun provideInteractor(tmdbWebService: TmdbWebService): MovieDetailsContract.Interactor {
+        return MovieDetailsInteractor(tmdbWebService)
     }
 
     @Provides
     @DetailsScope
-    fun providePresenter(detailsInteractor: MovieDetailsInteractor,
-                                  favoritesInteractor: FavoritesInteractor): MovieDetailsPresenter {
-        return MovieDetailsPresenterImpl(detailsInteractor, favoritesInteractor)
+    fun providePresenter(detailsInteractor: MovieDetailsContract.Interactor,
+                         favoritesInteractor: FavoritesContract.Interactor): MovieDetailsContract.Presenter {
+        return MovieDetailsPresenter(detailsInteractor, favoritesInteractor)
     }
 }
