+++ app/src/main/java/com/esoxjem/movieguide/movies/favorites/FavoritesInteractor.kt
@@ -5,9 +5,13 @@ import com.esoxjem.movieguide.movies.entities.Movie
 /**
  * @author arunsasidharan
  */
-interface FavoritesInteractor {
-    fun getFavorites(): List<Movie>
-    fun setFavorite(movie: Movie)
-    fun isFavorite(id: String): Boolean
-    fun unFavorite(id: String)
+internal class FavoritesInteractor(private val favoritesStore: FavoritesStore) : FavoritesContract.Interactor {
+
+    override fun getFavorites(): List<Movie> = favoritesStore.getFavorites()
+
+    override fun setFavorite(movie: Movie) = favoritesStore.setFavorite(movie)
+
+    override fun isFavorite(id: String): Boolean = favoritesStore.isFavorite(id)
+
+    override fun unFavorite(id: String) = favoritesStore.unfavorite(id)
 }
