--- app/src/main/java/com/esoxjem/movieguide/movies/favorites/FavoritesInteractorImpl.kt
@@ -1,17 +1,0 @@
-package com.esoxjem.movieguide.movies.favorites
-
-import com.esoxjem.movieguide.movies.entities.Movie
-
-/**
- * @author arunsasidharan
- */
-internal class FavoritesInteractorImpl(private val favoritesStore: FavoritesStore) : FavoritesInteractor {
-
-    override fun getFavorites(): List<Movie> = favoritesStore.getFavorites()
-
-    override fun setFavorite(movie: Movie) = favoritesStore.setFavorite(movie)
-
-    override fun isFavorite(id: String): Boolean = favoritesStore.isFavorite(id)
-
-    override fun unFavorite(id: String) = favoritesStore.unfavorite(id)
-}
