--- app/src/main/java/com/esoxjem/movieguide/movies/listing/MoviesListingView.kt
@@ -1,13 +1,0 @@
-package com.esoxjem.movieguide.movies.listing
-
-import com.esoxjem.movieguide.movies.entities.Movie
-
-/**
- * @author arun
- */
-interface MoviesListingView {
-    fun showMovies(movies: List<Movie>)
-    fun loadingStarted()
-    fun loadingFailed(errorMessage: String?)
-    fun onMovieClicked(movie: Movie)
-}
