--- app/src/main/java/com/esoxjem/movieguide/movies/sorting/SortingDialogView.kt
@@ -1,17 +1,0 @@
-package com.esoxjem.movieguide.movies.sorting
-
-/**
- * @author arunsasidharan
- */
-interface SortingDialogView {
-    fun setPopularChecked()
-
-    fun setNewestChecked()
-
-    fun setHighestRatedChecked()
-
-    fun setFavoritesChecked()
-
-    fun dismissDialog()
-
-}
