+++ app/src/test/java/com/esoxjem/movieguide/movies/details/MovieDetailsPresenterTest.java
@@ -1,9 +1,8 @@
 package com.esoxjem.movieguide.details;
 
 import com.esoxjem.movieguide.RxSchedulerRule;
-import com.esoxjem.movieguide.movies.details.MovieDetailsInteractor;
-import com.esoxjem.movieguide.movies.details.MovieDetailsPresenterImpl;
-import com.esoxjem.movieguide.movies.details.MovieDetailsView;
+import com.esoxjem.movieguide.movies.details.MovieDetailsContract;
+import com.esoxjem.movieguide.movies.details.MovieDetailsPresenter;
 import com.esoxjem.movieguide.movies.entities.Movie;
 import com.esoxjem.movieguide.movies.entities.Review;
 import com.esoxjem.movieguide.movies.entities.Video;
@@ -31,13 +30,13 @@ import static org.mockito.Mockito.when;
  * @author arunsasidharan
  */
 @RunWith(MockitoJUnitRunner.class)
-public class MovieDetailsPresenterImplTest {
+public class MovieDetailsPresenterTest {
     @Rule
     public RxSchedulerRule rule = new RxSchedulerRule();
     @Mock
-    private MovieDetailsView view;
+    private MovieDetailsContract.View view;
     @Mock
-    private MovieDetailsInteractor movieDetailsInteractor;
+    private MovieDetailsContract.Interactor movieDetailsInteractor;
     @Mock
     private FavoritesInteractor favoritesInteractor;
     @Mock
@@ -47,12 +46,12 @@ public class MovieDetailsPresenterImplTest {
     @Mock
     List<Review> reviews;
 
-    private MovieDetailsPresenterImpl movieDetailsPresenter;
+    private MovieDetailsPresenter movieDetailsPresenter;
 
     @Before
     public void setUp() throws Exception {
         MockitoAnnotations.initMocks(this);
-        movieDetailsPresenter = new MovieDetailsPresenterImpl(movieDetailsInteractor, favoritesInteractor);
+        movieDetailsPresenter = new MovieDetailsPresenter(movieDetailsInteractor, favoritesInteractor);
         movieDetailsPresenter.setView(view);
     }
